refactor abstract base
